shppingCartApplicationBackend/requirements.txt
djangorestframework
PyJWT==2.10.1
bcrypt==5.0.0
django-cors-headers==4.7.0